xconfadmin new refactoring

+++ adminapi/dcm/dcmformula_service.go
@@ -1,5 +1,5 @@
 /**
- * Copyright 2023 Comcast Cable Communications Management, LLC
+ * Copyright 2025 Comcast Cable Communications Management, LLC
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,26 +20,28 @@ package dcm
 import (
 	"errors"
 	"fmt"
-	"math"
 	"net/http"
 	"sort"
 	"strconv"
 	"strings"
-	"time"
+	"sync"
 
-	ru "xconfwebconfig/rulesengine"
+	xwcommon "github.com/rdkcentral/xconfwebconfig/common"
 
 	queries "xconfadmin/adminapi/queries"
+	"xconfadmin/common"
 	xcommon "xconfadmin/common"
 	xhttp "xconfadmin/http"
-	xutil "xconfadmin/util"
-	xwcommon "xconfwebconfig/common"
-	ds "xconfwebconfig/db"
-	xwhttp "xconfwebconfig/http"
-	"xconfwebconfig/rulesengine"
-	"xconfwebconfig/shared"
-	"xconfwebconfig/shared/logupload"
-	xwutil "xconfwebconfig/util"
+	core "xconfadmin/shared"
+
+	//	"xconfadmin/shared/logupload"
+	"xconfadmin/util"
+
+	ds "github.com/rdkcentral/xconfwebconfig/db"
+	xwhttp "github.com/rdkcentral/xconfwebconfig/http"
+	"github.com/rdkcentral/xconfwebconfig/rulesengine"
+	re "github.com/rdkcentral/xconfwebconfig/rulesengine"
+	"github.com/rdkcentral/xconfwebconfig/shared/logupload"
 
 	"github.com/google/uuid"
 )
@@ -49,8 +51,10 @@ const (
 	cDcmRulePageSize   = "pageSize"
 )
 
+var formulaUpdateMutex sync.Mutex
+
 func GetDcmFormulaAll() []*logupload.DCMGenericRule {
-	dcmformularules := logupload.GetDCMGenericRuleList()
+	dcmformularules := logupload.GetDCMGenericRuleListForAS()
 	return dcmformularules
 }
 
@@ -63,33 +67,45 @@ func GetDcmFormula(id string) *logupload.DCMGenericRule {
 
 }
 
-func validateUsageForDcmFormula(id string, appType string) (string, error) {
-	dcmformula := GetDcmFormula(id)
-	if dcmformula == nil || dcmformula.ApplicationType != appType {
-		return fmt.Sprintf("Entity with id  %s does not exist ", id), nil
+func validateIfExists(id string, appType string) error {
+	existingFormula := GetDcmFormula(id)
+	if existingFormula == nil || existingFormula.ApplicationType != appType {
+		return xwcommon.NewRemoteErrorAS(http.StatusNotFound, "Entity with id "+id+" does not exist ")
 	}
-	return "", nil
+	return nil
 }
 
-func DeleteDcmFormulabyId(id string, appType string) *xwhttp.ResponseEntity {
-	usage, err := validateUsageForDcmFormula(id, appType)
+func DeleteDcmFormulabyId(id string, appType string) *xcommon.ResponseEntity {
+	formulaUpdateMutex.Lock()
+	defer formulaUpdateMutex.Unlock()
+	err := validateIfExists(id, appType)
 	if err != nil {
-		return xwhttp.NewResponseEntity(http.StatusNotFound, err, nil)
-	}
-
-	if usage != "" {
-		return xwhttp.NewResponseEntity(http.StatusNotFound, errors.New(usage), nil)
+		return xcommon.NewResponseEntityWithStatus(http.StatusNotFound, err, nil)
 	}
 
 	err = DeleteOneDcmFormula(id, appType)
 	if err != nil {
-		return xwhttp.NewResponseEntity(http.StatusInternalServerError, err, nil)
+		return xcommon.NewResponseEntityWithStatus(http.StatusInternalServerError, err, nil)
 	}
 
-	return xwhttp.NewResponseEntity(http.StatusNoContent, nil, nil)
+	return xcommon.NewResponseEntityWithStatus(http.StatusNoContent, nil, nil)
+}
+
+func SaveDcmRules(itemList []core.Prioritizable) error {
+	for _, item := range itemList {
+		rule := item.(*logupload.DCMGenericRule)
+		if err := ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, rule.GetID(), rule); err != nil {
+			return err
+		}
+	}
+	return nil
 }
 
 func DeleteOneDcmFormula(id string, appType string) error {
+	existingRule := logupload.GetOneDCMGenericRule(id)
+	if existingRule == nil {
+		return fmt.Errorf("Entity with id %s does not exist", id)
+	}
 	err := ds.GetCachedSimpleDao().DeleteOne(ds.TABLE_DCM_RULE, id)
 	if err != nil {
 		return err
@@ -116,29 +132,11 @@ func DeleteOneDcmFormula(id string, appType string) error {
 		}
 	}
 
-	return packPriorities(appType)
-}
-
-func packPriorities(appType string) error {
-	changedRules := []*logupload.DCMGenericRule{}
-	dfrules := GetDcmRulesByApplicationType(appType)
-	// sort by ascending priority
-	sort.Slice(dfrules, func(i, j int) bool {
-		return dfrules[i].Priority < dfrules[j].Priority
-	})
-	priority := 1
-	for _, item := range dfrules {
-		if item.Priority != priority {
-			item.Priority = priority
-			changedRules = append(changedRules, item)
-		}
-		priority++
-	}
-	// Now save all updated priorities
-	for _, dcmrule := range changedRules {
-		if err := ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, dcmrule.ID, dcmrule); err != nil {
-			return err
-		}
+	dcmRulesByAppType := GetDcmRulesByApplicationType(appType)
+	prioritizableRules := DcmRulesToPrioritizables(dcmRulesByAppType)
+	err = SaveDcmRules(queries.PackPriorities(prioritizableRules, existingRule))
+	if err != nil {
+		return err
 	}
 	return nil
 }
@@ -148,19 +146,19 @@ func dcmRuleValidate(dfrule *logupload.DCMGenericRule) *xwhttp.ResponseEntity {
 		return xwhttp.NewResponseEntity(http.StatusBadRequest, errors.New("DCM formula Rule should be specified"), nil)
 	}
 
-	if xwutil.IsBlank(dfrule.ID) {
+	if util.IsBlank(dfrule.ID) {
 		dfrule.ID = uuid.New().String()
 	}
-	if xwutil.IsBlank(dfrule.ApplicationType) {
+	if util.IsBlank(dfrule.ApplicationType) {
 		return xwhttp.NewResponseEntity(http.StatusBadRequest, errors.New("ApplicationType is empty"), nil)
 	}
 
-	if xwutil.IsBlank(dfrule.Name) {
+	if util.IsBlank(dfrule.Name) {
 		return xwhttp.NewResponseEntity(http.StatusBadRequest, errors.New("Name is empty"), nil)
 	}
 
 	if dfrule.GetRule() != nil {
-		ru.NormalizeConditions(dfrule.GetRule())
+		rulesengine.NormalizeConditions(dfrule.GetRule())
 	} else {
 		return xwhttp.NewResponseEntity(http.StatusBadRequest, errors.New("Condition is empty"), nil)
 	}
@@ -198,9 +196,9 @@ func dcmRuleValidate(dfrule *logupload.DCMGenericRule) *xwhttp.ResponseEntity {
 
 func validatePercentage(dfrule *logupload.DCMGenericRule) error {
 	p := dfrule.Percentage
-	p1 := dfrule.PercentageL1
-	p2 := dfrule.PercentageL2
-	p3 := dfrule.PercentageL3
+	p1, _ := dfrule.PercentageL1.Int64()
+	p2, _ := dfrule.PercentageL2.Int64()
+	p3, _ := dfrule.PercentageL3.Int64()
 	if int(p1) < 0 || int(p2) < 0 || int(p3) < 0 {
 		err := fmt.Errorf("Percentage must be in range from 0 to 100")
 		return err
@@ -214,50 +212,13 @@ func validatePercentage(dfrule *logupload.DCMGenericRule) error {
 	return nil
 }
 
-func getAlteredSubList(itemsList []*logupload.DCMGenericRule, oldPriority int, newPriority int) []*logupload.DCMGenericRule {
-	start := int(math.Min(float64(oldPriority), float64(newPriority))) - 1
-	end := int(math.Max(float64(oldPriority), float64(newPriority)))
-	return itemsList[start:end]
-}
-
-func reorganizePriorities(dfrules []*logupload.DCMGenericRule, oldpriority int, newpriority int) []*logupload.DCMGenericRule {
-	if newpriority < 1 || newpriority > len(dfrules) {
-		newpriority = len(dfrules)
+func DcmRulesToPrioritizables(dcmRules []*logupload.DCMGenericRule) []core.Prioritizable {
+	prioritizables := make([]core.Prioritizable, len(dcmRules))
+	for i, item := range dcmRules {
+		itemCopy := *item
+		prioritizables[i] = &itemCopy
 	}
-	dfrule := dfrules[oldpriority-1]
-	dfrule.Priority = newpriority
-
-	if oldpriority < newpriority {
-		for i := oldpriority; i <= newpriority-1; i++ {
-			elem := dfrules[i]
-			elem.Priority = i
-			dfrules[i-1] = elem
-		}
-	}
-
-	if oldpriority > newpriority {
-		for i := oldpriority - 2; i >= newpriority-1; i-- {
-			elem := dfrules[i]
-			elem.Priority = i + 2
-			dfrules[i+1] = elem
-		}
-	}
-
-	dfrules[newpriority-1] = dfrule
-
-	return getAlteredSubList(dfrules, oldpriority, newpriority)
-}
-
-func AddnewItemAndRepriortize(newdfrule *logupload.DCMGenericRule) []*logupload.DCMGenericRule {
-	dfrules := GetDcmRulesByApplicationType(newdfrule.ApplicationType)
-	// sort by ascending priority
-	sort.Slice(dfrules, func(i, j int) bool {
-		return dfrules[i].Priority < dfrules[j].Priority
-	})
-	dfrules = append(dfrules, newdfrule)
-	oldpriority := len(dfrules)
-	newpriority := newdfrule.Priority
-	return reorganizePriorities(dfrules, oldpriority, newpriority)
+	return prioritizables
 }
 
 func CreateDcmRule(dfrule *logupload.DCMGenericRule, appType string) *xwhttp.ResponseEntity {
@@ -271,10 +232,13 @@ func CreateDcmRule(dfrule *logupload.DCMGenericRule, appType string) *xwhttp.Res
 		return respEntity
 	}
 
-	list := AddnewItemAndRepriortize(dfrule)
-	for _, entry := range list {
-		entry.Updated = xwutil.GetTimestamp(time.Now().UTC())
-		if err := ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, entry.ID, entry); err != nil {
+	formulaUpdateMutex.Lock()
+	defer formulaUpdateMutex.Unlock()
+	dcmRulesByAppType := GetDcmRulesByApplicationType(dfrule.ApplicationType)
+	changedDcmRules := queries.AddNewPrioritizableAndReorganizePriorities(dfrule, DcmRulesToPrioritizables(dcmRulesByAppType))
+	for _, entry := range changedDcmRules {
+		entry.(*logupload.DCMGenericRule).Updated = util.GetTimestamp()
+		if err := ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, entry.GetID(), entry); err != nil {
 			return xwhttp.NewResponseEntity(http.StatusInternalServerError, err, nil)
 		}
 	}
@@ -292,65 +256,44 @@ func GetDcmRulesByApplicationType(applicationType string) []*logupload.DCMGeneri
 	return list
 }
 
-func UpdateItemAndRepriortize(newdfrule *logupload.DCMGenericRule, oldPriority, newPriority int) (result []*logupload.DCMGenericRule, err error) {
-	dfrules := GetDcmRulesByApplicationType(newdfrule.ApplicationType)
-	// sort by ascending priority
-	sort.Slice(dfrules, func(i, j int) bool {
-		return dfrules[i].Priority < dfrules[j].Priority
-	})
-
-	if newPriority < 1 || newPriority > len(dfrules)+1 {
-		return nil, errors.New("Invalid value for Priority")
-	}
-	result = reorganizePriorities(dfrules, oldPriority, newPriority)
-
-	if len(result) > (newPriority - 1) {
-		result[newPriority-1] = newdfrule
-	} else if len(result) == 1 && oldPriority == newPriority {
-		result[0] = newdfrule
-	}
-
-	return result, nil
-}
-
-func UpdateDcmRule(dfrule *logupload.DCMGenericRule, appType string) *xwhttp.ResponseEntity {
-	if xwutil.IsBlank(dfrule.ID) {
+func UpdateDcmRule(incomingFormula *logupload.DCMGenericRule, appType string) *xwhttp.ResponseEntity {
+	if util.IsBlank(incomingFormula.ID) {
 		return xwhttp.NewResponseEntity(http.StatusBadRequest, errors.New("ID is empty"), nil)
 	}
-	if dfrule.ApplicationType != appType {
-		return xwhttp.NewResponseEntity(http.StatusConflict, fmt.Errorf("Entity with id %s ApplicationType doesn't match", dfrule.ID), nil)
+	if incomingFormula.ApplicationType != appType {
+		return xwhttp.NewResponseEntity(http.StatusConflict, fmt.Errorf("Entity with id %s ApplicationType doesn't match", incomingFormula.ID), nil)
 	}
-	existingRule := logupload.GetOneDCMGenericRule(dfrule.ID)
-	if existingRule == nil {
-		return xwhttp.NewResponseEntity(http.StatusConflict, fmt.Errorf("Entity with id %s does not exist", dfrule.ID), nil)
+	formulaUpdateMutex.Lock()
+	defer formulaUpdateMutex.Unlock()
+	existingFormula := logupload.GetOneDCMGenericRule(incomingFormula.ID)
+	if existingFormula == nil {
+		return xwhttp.NewResponseEntity(http.StatusConflict, fmt.Errorf("Entity with id %s does not exist", incomingFormula.ID), nil)
 	}
-	if existingRule.ApplicationType != dfrule.ApplicationType {
-		return xwhttp.NewResponseEntity(http.StatusConflict, fmt.Errorf("ApplicationType in db %s doesn't match the ApplicationType %s in req", existingRule.ApplicationType, dfrule.ApplicationType), nil)
+	if existingFormula.ApplicationType != incomingFormula.ApplicationType {
+		return xwhttp.NewResponseEntity(http.StatusConflict, fmt.Errorf("ApplicationType in db %s doesn't match the ApplicationType %s in req", existingFormula.ApplicationType, incomingFormula.ApplicationType), nil)
 	}
-	respEntity := dcmRuleValidate(dfrule)
+	respEntity := dcmRuleValidate(incomingFormula)
 	if respEntity.Error != nil {
 		return respEntity
 	}
 
-	if dfrule.Priority == existingRule.Priority {
-		dfrule.Updated = xwutil.GetTimestamp(time.Now().UTC())
-		if err := ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, dfrule.ID, dfrule); err != nil {
+	if incomingFormula.Priority == existingFormula.Priority {
+		incomingFormula.Updated = util.GetTimestamp()
+		if err := ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, incomingFormula.ID, incomingFormula); err != nil {
 			return xwhttp.NewResponseEntity(http.StatusInternalServerError, err, nil)
 		}
 	} else {
-		list, err := UpdateItemAndRepriortize(dfrule, existingRule.Priority, dfrule.Priority)
-		if err != nil {
-			return xwhttp.NewResponseEntity(http.StatusBadRequest, err, nil)
-		}
-		for _, entry := range list {
-			entry.Updated = xwutil.GetTimestamp(time.Now().UTC())
-			if err = ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, entry.ID, entry); err != nil {
+		formulasByApplicationType := GetDcmRulesByApplicationType(incomingFormula.ApplicationType)
+		changedFormulae := queries.UpdatePrioritizablePriorityAndReorganize(incomingFormula, DcmRulesToPrioritizables(formulasByApplicationType), existingFormula.Priority)
+		for _, entry := range changedFormulae {
+			entry.(*logupload.DCMGenericRule).Updated = util.GetTimestamp()
+			if err := ds.GetCachedSimpleDao().SetOne(ds.TABLE_DCM_RULE, entry.GetID(), entry); err != nil {
 				return xwhttp.NewResponseEntity(http.StatusInternalServerError, err, nil)
 			}
 		}
 	}
 
-	return xwhttp.NewResponseEntity(http.StatusOK, nil, dfrule)
+	return xwhttp.NewResponseEntity(http.StatusOK, nil, incomingFormula)
 }
 
 func dcmRuleGeneratePage(list []*logupload.DCMGenericRule, page int, pageSize int) (result []*logupload.DCMGenericRule) {
@@ -388,57 +331,29 @@ func DcmFormulaRuleGeneratePageWithContext(dfrules []*logupload.DCMGenericRule, 
 }
 
 func DcmFormulaFilterByContext(searchContext map[string]string) []*logupload.DCMGenericRule {
-	dcmFormulaRules := logupload.GetDCMGenericRuleList()
+	dcmFormulaRules := logupload.GetDCMGenericRuleListForAS()
 	dcmFormulaRuleList := []*logupload.DCMGenericRule{}
 	for _, dcmRule := range dcmFormulaRules {
 		if dcmRule == nil {
 			continue
 		}
-		if applicationType, ok := xutil.FindEntryInContext(searchContext, xwcommon.APPLICATION_TYPE, false); ok {
-			if dcmRule.ApplicationType != applicationType && dcmRule.ApplicationType != shared.ALL {
+		if applicationType, ok := util.FindEntryInContext(searchContext, core.APPLICATION_TYPE, false); ok {
+			if dcmRule.ApplicationType != applicationType && dcmRule.ApplicationType != core.ALL {
 				continue
 			}
 		}
-		if name, ok := xutil.FindEntryInContext(searchContext, xcommon.NAME_UPPER, false); ok {
+		if name, ok := util.FindEntryInContext(searchContext, xcommon.NAME_UPPER, false); ok {
 			if !strings.Contains(strings.ToLower(dcmRule.Name), strings.ToLower(name)) {
 				continue
 			}
 		}
-		if key, ok := xutil.FindEntryInContext(searchContext, xcommon.FREE_ARG, false); ok {
-			keyMatch := false
-			for _, condition := range ru.ToConditions(dcmRule.GetRule()) {
-				if strings.Contains(strings.ToLower(condition.GetFreeArg().Name), strings.ToLower(key)) {
-					keyMatch = true
-					break
-				}
-			}
-			if !keyMatch {
+		if key, ok := util.FindEntryInContext(searchContext, xcommon.FREE_ARG, false); ok {
+			if !re.IsExistConditionByFreeArgName(*dcmRule.GetRule(), key) {
 				continue
 			}
 		}
-		if fixedArgValue, ok := xutil.FindEntryInContext(searchContext, xcommon.FIXED_ARG, false); ok {
-			valueMatch := false
-			for _, condition := range ru.ToConditions(dcmRule.GetRule()) {
-				if condition.GetFixedArg() != nil && condition.GetFixedArg().IsCollectionValue() {
-					fixedArgs := condition.GetFixedArg().GetValue().([]string)
-					for _, fixedArg := range fixedArgs {
-						if strings.Contains(strings.ToLower(fixedArg), strings.ToLower(fixedArgValue)) {
-							valueMatch = true
-							break
-						}
-					}
-				}
-				if valueMatch {
-					break
-				}
-				if condition.GetOperation() != rulesengine.StandardOperationExists && condition.GetFixedArg() != nil && condition.GetFixedArg().IsStringValue() {
-					if strings.Contains(strings.ToLower(condition.FixedArg.Bean.Value.JLString), strings.ToLower(fixedArgValue)) {
-						valueMatch = true
-						break
-					}
-				}
-			}
-			if !valueMatch {
+		if fixedArgValue, ok := util.FindEntryInContext(searchContext, xcommon.FIXED_ARG, false); ok {
+			if !re.IsExistConditionByFixedArgValue(*dcmRule.GetRule(), fixedArgValue) {
 				continue
 			}
 		}
@@ -453,17 +368,17 @@ func importFormula(formulaWithSettings *logupload.FormulaWithSettings, overwrite
 	logUploadSettings := formulaWithSettings.LogUpLoadSettings
 	vodSettings := formulaWithSettings.VodSettings
 
-	if xwutil.IsBlank(formula.ApplicationType) {
+	if util.IsBlank(formula.ApplicationType) {
 		formula.ApplicationType = appType
 	}
 	if deviceSettings != nil {
-		if xwutil.IsBlank(deviceSettings.ApplicationType) {
+		if util.IsBlank(deviceSettings.ApplicationType) {
 			deviceSettings.ApplicationType = appType
 		}
 		if formula.ApplicationType != deviceSettings.ApplicationType {
 			return xwhttp.NewResponseEntity(http.StatusBadRequest, errors.New("DeviceSettings ApplicationType mismatch"), nil)
 		}
-		if xwutil.IsBlank(deviceSettings.Schedule.TimeZone) {
+		if util.IsBlank(deviceSettings.Schedule.TimeZone) {
 			if logUploadSettings != nil {
 				logUploadSettings.Schedule.TimeZone = logupload.UTC
 			}
@@ -473,13 +388,13 @@ func importFormula(formulaWithSettings *logupload.FormulaWithSettings, overwrite
 		}
 	}
 	if logUploadSettings != nil {
-		if xwutil.IsBlank(logUploadSettings.ApplicationType) {
+		if util.IsBlank(logUploadSettings.ApplicationType) {
 			logUploadSettings.ApplicationType = appType
 		}
 		if formula.ApplicationType != logUploadSettings.ApplicationType {
 			return xwhttp.NewResponseEntity(http.StatusBadRequest, errors.New("logUploadSettings ApplicationType mismatch"), nil)
 		}
-		if xwutil.IsBlank(logUploadSettings.Schedule.TimeZone) {
+		if util.IsBlank(logUploadSettings.Schedule.TimeZone) {
 			logUploadSettings.Schedule.TimeZone = logupload.UTC
 		}
 		if respEntity := LogUploadSettingsValidate(logUploadSettings); respEntity.Error != nil {
@@ -487,7 +402,7 @@ func importFormula(formulaWithSettings *logupload.FormulaWithSettings, overwrite
 		}
 	}
 	if vodSettings != nil {
-		if xwutil.IsBlank(vodSettings.ApplicationType) {
+		if util.IsBlank(vodSettings.ApplicationType) {
 			vodSettings.ApplicationType = appType
 		}
 		if formula.ApplicationType != vodSettings.ApplicationType {
@@ -553,13 +468,13 @@ func importFormulas(formulaWithSettingsList []*logupload.FormulaWithSettings, ap
 		respEntity := importFormula(formulaWithSettings, overwrite, appType)
 		if respEntity.Error != nil {
 			entityMessage := xhttp.EntityMessage{
-				Status:  xcommon.ENTITY_STATUS_FAILURE,
+				Status:  common.ENTITY_STATUS_FAILURE,
 				Message: respEntity.Error.Error(),
 			}
 			entitiesMap[formula.ID] = entityMessage
 		} else {
 			entityMessage := xhttp.EntityMessage{
-				Status:  xcommon.ENTITY_STATUS_SUCCESS,
+				Status:  common.ENTITY_STATUS_SUCCESS,
 				Message: formula.ID,
 			}
 			entitiesMap[formula.ID] = entityMessage
